Added chunk.go by refactoring
Improved chunk management.
Centralize adding and removing chunks.
